refactoring: move LazyChunkedRemoteFile into API_client module
